+++ react essentials/01-starting-project/src/App.jsx
@@ -1,57 +1,18 @@
-import { useState } from "react";
+import { Fragment } from "react";
 import Header from "./components/Header/Header.jsx";
-import CoreConcept from "./components/CoreConcept.jsx";
-import TabButton from "./components/TabButton.jsx";
-
-import { CORE_CONCEPTS } from "./data.js";
-import { EXAMPLES } from "./data.js";
+import CoreConcepts from "./components/CoreConcepts.jsx";
+import Examples from "./components/Examples.jsx";
 
 function App() {
-  const [selectedTopic, setSelectedTopic] = useState();
 
-  function selectHandler(selectedButton) {
-    setSelectedTopic(selectedButton);
-  }
   return (
-    <div>
+    <Fragment>
       <Header />
       <main>
-        <section id="core-concepts">
-          <h2>Core Concepts</h2>
-          <ul>
-            {/* <CoreConcept{...CORE_CONCEPTS[0]}/>
-            <CoreConcept{...CORE_CONCEPTS[1]}/> */}
-
-            {CORE_CONCEPTS.map((concept, i) => (
-              <CoreConcept key={i}
-              {...concept}
-              />
-            ))}
-          </ul>
-        </section>
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton isSelected={selectedTopic === "components"} onSelect={() => selectHandler("components")}>
-              Components
-            </TabButton>
-            <TabButton isSelected={selectedTopic === "jsx"} onSelect={() => selectHandler("jsx")}>JSX</TabButton>
-            <TabButton isSelected={selectedTopic === "props"}  onSelect={() => selectHandler("props")}>Props</TabButton>
-            <TabButton isSelected={selectedTopic === "state"}  onSelect={() => selectHandler("state")}>State</TabButton>
-          </menu>
-          {!selectedTopic && <p>please select a topic</p>}
-          {selectedTopic && (
-            <div id="tab-content">
-              <h3>{EXAMPLES[selectedTopic].title}</h3>
-              <p>{EXAMPLES[selectedTopic].description}</p>
-              <pre>
-                <code>{EXAMPLES[selectedTopic].code}</code>
-              </pre>
-            </div>
-          )}
-        </section>
+        <CoreConcepts />
+        <Examples />
       </main>
-    </div>
+    </Fragment>
   );
 }
